requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
Flask-Login
